# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Smart Building API
  description: Sistema de controle inteligente de ar-condicionado
  version: 1.0.0
paths:
...
  /api/v1/rooms:
    get:
      tags:
        - Rooms
      summary: Listar salas
      operationId: list_rooms_api_v1_rooms_get
      security:
        - HTTPBearer: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Size
        - name: building
          in: query
          required: false
          schema:
            anyOf:
              - type: string
                maxLength: 100
              - type: 'null'
            title: Building
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response List Rooms Api V1 Rooms Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - Rooms
      summary: Cadastrar sala
      operationId: create_room_api_v1_rooms_post
      security:
        - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  '/api/v1/rooms/{room_id}':
    get:
      tags:
        - Rooms
      summary: Detalhes de uma sala
      operationId: get_room_api_v1_rooms__room_id__get
      security:
        - HTTPBearer: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Room Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
        - Rooms
      summary: Atualizar sala
      operationId: update_room_api_v1_rooms__room_id__patch
      security:
        - HTTPBearer: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Room Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
        - Rooms
      summary: Excluir sala
      operationId: delete_room_api_v1_rooms__room_id__delete
      security:
        - HTTPBearer: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: Room Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/alerts:
    get:
      tags:
        - Alerts
      summary: Listar alertas
      operationId: list_alerts_api_v1_alerts_get
      security:
        - HTTPBearer: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            title: Page
        - name: size
          in: query
          required: false
          schema:
            type: integer
            maximum: 100
            minimum: 1
            default: 20
            title: Size
        - name: active_only
          in: query
          required: false
          schema:
            type: boolean
            description: Somente alertas não resolvidos
            default: false
            title: Active Only
          description: Somente alertas não resolvidos
        - name: device_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
                format: uuid
              - type: 'null'
            title: Device Id
        - name: severity
          in: query
          required: false
          schema:
            anyOf:
              - $ref: '#/components/schemas/AlertSeverity'
              - type: 'null'
            title: Severity
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response List Alerts Api V1 Alerts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
        - Alerts
      summary: Criar alerta manualmente
      operationId: create_alert_api_v1_alerts_post
      security:
        - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/consumption:
    get:
      tags:
        - Consumption
      summary: Histórico de consumo energético
      operationId: get_consumption_api_v1_consumption_get
      security:
        - HTTPBearer: []
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            pattern: ^(24h|7d|30d)$
            default: 24h
            title: Period
        - name: device_id
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Device Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Consumption Api V1 Consumption Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/reports/consumption:
    get:
      tags:
        - Reports
      summary: Relatório de consumo
      operationId: get_consumption_report_api_v1_reports_consumption_get
      security:
        - HTTPBearer: []
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            pattern: ^(24h|7d|30d)$
            default: 7d
            title: Period
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Consumption Report Api V1 Reports Consumption Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reports/consumption/export:
    get:
      tags:
        - Reports
      summary: Exportar relatório de consumo em CSV
      description: Exporta o relatório de consumo como arquivo CSV.
      operationId: export_consumption_csv_api_v1_reports_consumption_export_get
      security:
        - HTTPBearer: []
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            pattern: ^(24h|7d|30d)$
            default: 7d
            title: Period
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/reports/alerts:
    get:
      tags:
        - Reports
      summary: Relatório de alertas
      operationId: get_alerts_report_api_v1_reports_alerts_get
      security:
        - HTTPBearer: []
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            pattern: ^(24h|7d|30d)$
            default: 7d
            title: Period
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Get Alerts Report Api V1 Reports Alerts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AlertCreate:
      properties:
        device_id:
          type: string
          format: uuid
          title: Device Id
        alert_type:
          $ref: '#/components/schemas/AlertType'
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        message:
          anyOf:
            - type: string
            - type: 'null'
          title: Message
      type: object
      required:
        - device_id
        - alert_type
        - severity
      title: AlertCreate
    AlertResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        device_id:
          type: string
          format: uuid
          title: Device Id
        alert_type:
          $ref: '#/components/schemas/AlertType'
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        message:
          anyOf:
            - type: string
            - type: 'null'
          title: Message
        created_at:
          type: string
          format: date-time
          title: Created At
        acknowledged_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Acknowledged At
        resolved_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Resolved At
        is_active:
          type: boolean
          title: Is Active
          default: true
        links:
          items:
            $ref: '#/components/schemas/HateoasLink'
          type: array
          title: Links
      type: object
      required:
        - id
        - device_id
        - alert_type
        - severity
        - created_at
      title: AlertResponse
    AlertSeverity:
      type: string
      enum:
        - info
        - warning
        - critical
      title: AlertSeverity
    AlertType:
      type: string
      enum:
        - HIGH_TEMPERATURE
        - LOW_HUMIDITY
        - ANOMALY_DETECTED
        - DEVICE_OFFLINE
        - WINDOW_OPEN
        - CONSUMPTION_LIMIT
      title: AlertType
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HateoasLink:
      properties:
        rel:
          type: string
          title: Rel
        href:
          type: string
          title: Href
        method:
          type: string
          title: Method
          default: GET
      type: object
      required:
        - rel
        - href
      title: HateoasLink
...
    RoomCreate:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        building:
          type: string
          maxLength: 100
          minLength: 1
          title: Building
        floor:
          type: integer
          maximum: 100
          minimum: 0
          title: Floor
        area_m2:
          anyOf:
            - type: number
              exclusiveMinimum: 0
            - type: 'null'
          title: Area M2
      type: object
      required:
        - name
        - building
        - floor
      title: RoomCreate
    RoomResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        building:
          type: string
          title: Building
        floor:
          type: integer
          title: Floor
        area_m2:
          anyOf:
            - type: number
            - type: 'null'
          title: Area M2
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Updated At
        links:
          items:
            $ref: '#/components/schemas/HateoasLink'
          type: array
          title: Links
      type: object
      required:
        - id
        - name
        - building
        - floor
        - created_at
      title: RoomResponse
    RoomUpdate:
      properties:
        name:
          anyOf:
            - type: string
              maxLength: 100
              minLength: 1
            - type: 'null'
          title: Name
        building:
          anyOf:
            - type: string
              maxLength: 100
              minLength: 1
            - type: 'null'
          title: Building
        floor:
          anyOf:
            - type: integer
              maximum: 100
              minimum: 0
            - type: 'null'
          title: Floor
        area_m2:
          anyOf:
            - type: number
              exclusiveMinimum: 0
            - type: 'null'
          title: Area M2
      type: object
      title: RoomUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...